Rename .current.js config file to current.js
