+++ decentraland/.gitignore
@@ -5,4 +5,4 @@ npm-debug.*
 *.p12
 *.key
 *.mobileprovision
-config/.current.js
+config/current.js
